Refactors login service and tests' nomenclature
