+++ bestofbooks/src/test/java/tqs/group4/bestofbooks/service/LoginServiceTest.java
@@ -21,6 +21,7 @@ import tqs.group4.bestofbooks.exception.RegistrationFailedException;
 import tqs.group4.bestofbooks.exception.RepeatedPublisherNameException;
 import tqs.group4.bestofbooks.exception.RepeatedUsernameException;
 import tqs.group4.bestofbooks.exception.UserNotFoundException;
+import tqs.group4.bestofbooks.mocks.BuyerMock;
 import tqs.group4.bestofbooks.mocks.OrderMocks;
 import tqs.group4.bestofbooks.mocks.PublisherMocks;
 import tqs.group4.bestofbooks.model.Admin;
@@ -33,9 +34,7 @@ import tqs.group4.bestofbooks.repository.PublisherRepository;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
-import java.util.Optional;
-
-public class LoginServicesTest {
+public class LoginServiceTest {
 	
 	@Mock
 	private AdminRepository adminRepository;
@@ -47,7 +46,7 @@ public class LoginServicesTest {
 	private BuyerRepository buyerRepository;
 	
 	@InjectMocks
-	private LoginServices loginService;
+	private LoginService loginService;
 
 	private OrderDTO orderDTO = OrderDTO.fromOrder(OrderMocks.order1);
 	
@@ -498,6 +497,20 @@ public class LoginServicesTest {
 		when(publisherRepository.existsByUsername("pub1")).thenReturn(true);
 		when(publisherRepository.findByUsername("pub1")).thenReturn(Optional.of(PublisherMocks.publisher1));
 		assertDoesNotThrow(() -> loginService.checkIfUserIsTheRightPublisher("Publisher 1", "pub1"));
+	}
+
+	@Test
+	void searchBuyerByExistentUsername_getBuyer() throws UserNotFoundException {
+		String existentUsername = BuyerMock.buyer1.getUsername();
+		when(buyerRepository.findById(existentUsername)).thenReturn(java.util.Optional.of(BuyerMock.buyer1));
+		assertEquals(BuyerMock.buyer1, loginService.getBuyerFromUsername(existentUsername));
+	}
+
+	@Test
+	void searchBuyerByUnknownIsbn_UserNotFoundExceptionShouldBeThrown() {
+		String unknownUsername = "None";
+		assertThrows(UserNotFoundException.class,
+				() -> loginService.getBuyerFromUsername(unknownUsername));
 	}
 
 }
